+++ parser/dex/terraswap/columbus_v1/app.go
@@ -28,7 +28,7 @@ func New(repo p_dex.PairRepo, logger logging.Logger, c configs.ParserConfig) (p_
 	}
 
 	parsers := p_dex.PairParsers{
-		CreatePairParser: parser.NewParser[p_dex.ParsedTx](finder, &createPairMapper{}),
+		CreatePairParser: parser.NewParser[p_dex.ParsedTx](finder, p_dex.NewFactoryMapper()),
 		PairActionParser: nil,
 		InitialProvide:   nil,
 		WasmTransfer:     nil,
@@ -115,12 +115,12 @@ func (p *terraswapApp) updateParsers(pairs map[string]p_dex.Pair) error {
 	if err != nil {
 		return errors.Wrap(err, "createParsers")
 	}
-	p.Parsers.WasmTransfer = parser.NewParser[p_dex.ParsedTx](wasmTransferFinder, &wasmCommonTransferMapper{pairSet: pairs})
+	p.Parsers.WasmTransfer = parser.NewParser[p_dex.ParsedTx](wasmTransferFinder, p_dex.NewWasmTransferMapper(dex.WasmTransferLegacyCw20AddrKey, pairs))
 
 	transferRule, err := cv1.CreateTransferRuleFinder(nil)
 	if err != nil {
 		return errors.Wrap(err, "createParsers")
 	}
-	p.Parsers.Transfer = parser.NewParser[p_dex.ParsedTx](transferRule, &transferMapper{pairSet: pairs})
+	p.Parsers.Transfer = parser.NewParser[p_dex.ParsedTx](transferRule, p_dex.NewTransferMapper(pairs))
 	return nil
 }
