+++ parser/dex/terraswap/columbus_v1/mappers.go
@@ -2,7 +2,6 @@ package columbus_v1
 
 import (
 	"fmt"
-	"strings"
 
 	"github.com/dezswap/cosmwasm-etl/parser"
 	"github.com/dezswap/cosmwasm-etl/parser/dex"
@@ -12,14 +11,10 @@ import (
 	"github.com/pkg/errors"
 )
 
-var _ parser.Mapper[dex.ParsedTx] = &createPairMapper{}
 var _ parser.Mapper[dex.ParsedTx] = &pairMapper{}
-var _ parser.Mapper[dex.ParsedTx] = &transferMapper{}
-var _ parser.Mapper[dex.ParsedTx] = &wasmCommonTransferMapper{}
 
 type mapperMixin struct{}
 
-type createPairMapper struct{ mapperMixin }
 type pairMapper struct {
 	mixin   mapperMixin
 	pairSet map[string]dex.Pair
@@ -34,29 +29,6 @@ type wasmCommonTransferMapper struct {
 }
 
 // match implements mapper
-func (m *createPairMapper) MatchedToParsedTx(res eventlog.MatchedResult, optionals ...interface{}) ([]*dex.ParsedTx, error) {
-	if err := m.mapperMixin.checkResult(res, columbus_v1.CreatePairMatchedLen); err != nil {
-		return nil, errors.Wrap(err, "createPairMapper.MatchedToParsedTx")
-	}
-
-	assets := strings.Split(res[columbus_v1.FactoryPairIdx].Value, "-")
-	if len(assets) != 2 {
-		msg := fmt.Sprintf("expected assets length(%d)", 2)
-		return nil, errors.New(msg)
-	}
-
-	return []*dex.ParsedTx{{
-		Type:         dex.CreatePair,
-		Sender:       "",
-		ContractAddr: res[columbus_v1.FactoryPairAddrIdx].Value,
-		Assets: [2]dex.Asset{
-			{Addr: assets[0]},
-			{Addr: assets[1]},
-		},
-		LpAddr:   res[columbus_v1.FactoryLpAddrIdx].Value,
-		LpAmount: "",
-	}}, nil
-}
 
 // match implements mapper
 func (m *pairMapper) MatchedToParsedTx(res eventlog.MatchedResult, optionals ...interface{}) ([]*dex.ParsedTx, error) {
@@ -170,113 +142,6 @@ func (m *pairMapper) withdrawMatchedToParsedTx(res eventlog.MatchedResult, pair 
 		},
 	}}, nil
 
-}
-
-// match implements mapper
-func (m *wasmCommonTransferMapper) MatchedToParsedTx(res eventlog.MatchedResult, optionals ...interface{}) ([]*dex.ParsedTx, error) {
-	if err := m.mixin.checkResult(res); err != nil {
-		return nil, errors.Wrap(err, "wasmCommonTransferMapper.MatchedToParsedTx")
-	}
-
-	fp, fromPair := m.pairSet[res[columbus_v1.WasmTransferFromIdx].Value]
-	tp, toPair := m.pairSet[res[columbus_v1.WasmTransferToIdx].Value]
-	if fromPair && toPair {
-		msg := fmt.Sprintf("cannot be both from and to, see the tx, result(%v)", res)
-		return nil, errors.New(msg)
-	}
-
-	if !fromPair && !toPair {
-		return nil, nil
-	}
-
-	pair := fp
-	if toPair {
-		pair = tp
-	}
-
-	assets := [2]dex.Asset{
-		{Addr: pair.Assets[0]},
-		{Addr: pair.Assets[1]},
-	}
-	meta := make(map[string]interface{})
-
-	target := res[columbus_v1.WasmTransferCw20AddrIdx].Value
-	idx := dex.IndexOf(pair.Assets, target)
-	if idx == -1 {
-		meta[target] = res[columbus_v1.WasmTransferAmountIdx].Value
-	} else {
-		for _, item := range res {
-			if item.Key == "amount" {
-				assets[idx].Amount = item.Value
-			}
-		}
-	}
-
-	if fromPair {
-		assets[idx].Amount = fmt.Sprintf("-%s", assets[idx].Amount)
-	}
-
-	return []*dex.ParsedTx{{
-		Type:         dex.Transfer,
-		Sender:       res[columbus_v1.WasmTransferFromIdx].Value,
-		ContractAddr: pair.ContractAddr,
-		Assets:       assets,
-		Meta:         meta,
-	}}, nil
-
-}
-
-// match implements mapper
-func (m *transferMapper) MatchedToParsedTx(res eventlog.MatchedResult, optionals ...interface{}) ([]*dex.ParsedTx, error) {
-	fp, fromPair := m.pairSet[res[columbus_v1.TransferSenderIdx].Value]
-	tp, toPair := m.pairSet[res[columbus_v1.TransferRecipientIdx].Value]
-
-	if fromPair && toPair {
-		msg := fmt.Sprintf("cannot be both from and to, see the tx, result(%v)", res)
-		return nil, errors.New(msg)
-	}
-
-	if !fromPair && !toPair {
-		return nil, nil
-	}
-
-	pair := fp
-	if toPair {
-		pair = tp
-	}
-
-	assets := [2]dex.Asset{
-		{Addr: pair.Assets[0]},
-		{Addr: pair.Assets[1]},
-	}
-	meta := make(map[string]interface{})
-
-	amountStrs := strings.Split(res[columbus_v1.TransferAmountIdx].Value, ",")
-	for _, amountStr := range amountStrs {
-		asset, err := dex.GetAssetFromAmountAssetString(amountStr)
-		if err != nil {
-			return nil, errors.Wrap(err, "transferMapper.MatchedToParsedTx")
-		}
-		idx := dex.IndexOf(pair.Assets, asset.Addr)
-		if idx == -1 {
-			meta[asset.Addr] = asset.Amount
-		} else {
-			assets[idx] = asset
-		}
-		if fromPair {
-			assets[idx].Amount = fmt.Sprintf("-%s", assets[idx].Amount)
-		}
-	}
-
-	return []*dex.ParsedTx{{
-		Type:         dex.Transfer,
-		Sender:       res[columbus_v1.TransferSenderIdx].Value,
-		ContractAddr: pair.ContractAddr,
-		Assets:       assets,
-		LpAddr:       "",
-		LpAmount:     "",
-		Meta:         meta,
-	}}, nil
 }
 
 func (*mapperMixin) checkResult(res eventlog.MatchedResult, expectedLen ...int) error {
