+++ pkg/dex/terraswap/columbus_v1/consts.go
@@ -13,7 +13,6 @@ const (
 )
 
 const (
-	CreatePairMatchedLen   = FactoryLpAddrIdx + 1
 	PairSwapMatchedLen     = PairSwapCommissionAmountIdx + 1
 	PairProvideMatchedLen  = PairProvideShareIdx + 1
 	PairWithdrawMatchedLen = PairWithdrawRefundAssetsIdx + 1
@@ -23,10 +22,6 @@ const (
 
 const (
 	FactoryAddrIdx = iota
-	FactoryActionIdx
-	FactoryPairIdx
-	FactoryPairAddrIdx
-	FactoryLpAddrIdx
 )
 
 const (
